requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
crispy-bootstrap5==2024.2
cryptography==43.0.0
cssselect2==0.7.0
Django==5.1
django-crispy-forms==2.3
django-filter==24.3
django-jalali==7.0.0
et-xmlfile==1.1.0
fonttools==4.53.1
html5lib==1.1
idna==3.7
jalali_core==1.0.0
jdatetime==5.0.0
lxml==5.3.0
numpy==2.1.0
openpyxl==3.1.5
oscrypto==1.3.0
pandas==2.2.2
pillow==10.4.0
pycparser==2.22
pydyf==0.11.0
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
pypdf==4.3.1
pyphen==0.16.0
pypng==0.20220715.0
python-bidi==0.6.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.2
requests==2.32.3
six==1.16.0
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.3.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
Unidecode==1.3.8
uritools==4.0.3
urllib3==2.2.2
weasyprint==62.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zopfli==0.2.3
